tweaks to output
First, a bit of reordering. Instead of:

    FAILURE test_whatever:
    | >       assert "yes" == "no"
    | E       AssertionError: assert 'yes' == 'no'
    | E         - no
    | E         + yes
    | test_file.py:6: AssertionError
    ::error ...

...move the "::error" to right under "FAILURE":

    FAILURE test_whatever:
    ::error ...
    | >       assert "yes" == "no"
    | E       AssertionError: assert 'yes' == 'no'
    | E         - no
    | E         + yes
    | test_file.py:6: AssertionError

That makes it easier to visually group them (without it, it's ambiguous
    whether the "::error" goes with this failure, or the next one).

Next, tweak the "::error" line. We currently provide the file path and
line number, but GH really wants those to be for files that changed in
the most previous commit. It auto-generates a link, but that link is to
that file/line within the diff for most recent change — not the full
file. If the file didn't change in that commit (which will be common for
these integ tests), the link is useless; and the fact that it's a
link makes it harder to select just a portion, for copy-pasting. So, get
rid of those in the formal parameters of the "::error" command, and
instead just put them into the message (where they'll get rendered as
plaintext).